The expected total number of CFUs in the mixed sample in the multiple mixing schemes at each stage of the mixing process.

This function calculates the resulting expected total number of CFUs in the mixed sample in the multiple mixing plans at each stage of the mixing process.

sim_multiple_stages(mu, sigma, alpha_in, k, l, r, distribution, n_sim)

Arguments

mu

the average number of CFUs (\(\mu\)) in the mixed sample, which is in a logarithmic scale if we use a Lognormal / Poisson lognormal distribution

sigma

the standard deviation of the colony-forming units in the mixed sample on the logarithmic scale (default value 0.8)

alpha_in

concentration parameter at the initial stage

k

number of small portions / primary samples

l

number of revolutions / stages

r

the rate of the concentration parameter changes at each mixing stage

distribution

what suitable distribution type we have employed for simulation such as "Poisson-Type A" or "Poisson-Type B" or "Lognormal-Type A" or "Lognormal-Type B" or "Poisson lognormal-Type A" or "Poisson lognormal-Type B"

n_sim

number of simulations

Value

The expected total number of CFUs in each revolution / stage.

Details

Let \(N'\) be the number of CFUs in the mixed sample, which is produced by the mixing of \(k\) primary samples and \(N' = \sum N_i\) and let \(N_i\) be the number of CFUs. For this package development, we have employed the notations 'Type-A' and 'Type-B' to indicate the type of distributions, which are applied in the previous literature as 'fair' and 'beta', respectively; see Nauta (2005).

This package will consider stage-by-stage the mixing process and assumes systematically breaking clusters at every stage of the mixing. Therefore, it can be assumed the concentration parameter also systematically changes with the concentration of the contribution.

References

Nauta, M.J., 2005. Microbiological risk assessment models for partitioning and mixing during food handling. International Journal of Food Microbiology 100, 311-322.

See also

Examples

mu <- 100
sigma <- 0.8
alpha_in <- 0.01
k <- c(30,75)
l <- 250
rate <- 0.01
distribution <- c("Poisson lognormal-Type B","Poisson lognormal-Type B")
n_sim <- 2000
colMeans(sim_multiple_stages(mu, sigma, alpha_in, k, l, rate, distribution, n_sim))
#> mixing plan (mu = 100.0, k = 30, Poisson lognormal-Type B) 
#>                                                    137.652 
#> mixing plan (mu = 100.0, k = 75, Poisson lognormal-Type B) 
#>                                                    137.672