The generated number of colony-forming units in the mixed sample by the simulation results in the single mixing plan with a single stage of the mixing.

This function calculates the resulting generated number of colony forming units in the mixed sample in the single mixing plan with single stage of the mixing.

sim_single(mu, sigma, alpha, k, distribution, n_sim, summary = 1)

Arguments

mu
the average number of CFUs (\(\mu\)) in the mixed sample, which is in a logarithmic scale if we use a Lognormal / Poisson lognormal distribution

sigma
the standard deviation of the colony-forming units in the mixed sample on the logarithmic scale (default value 0.8)

alpha
concentration parameter

k
number of small portions / primary samples

distribution
what suitable distribution type we have employed for simulation such as "Poisson-Type A" or "Poisson-Type B" or "Lognormal-Type A" or "Lognormal-Type B" or "Poisson lognormal-Type A" or "Poisson lognormal-Type B"

n_sim
number of simulations

summary
need to select one number from the list 1 to 5 which depends on what simulated observations are needed (default 1). If summary = 1, the function provides the expected total number of the simulated \(N'\) after n_sim simulations are run. If summary = 2, the function provides generated CFUs in each primary sample after n_sim simulations are run. If summary = 3, the function provides the expected total number of the simulated \(N'\) at each simulation run. If summary = 4, the function provides the mean and variance values of the simulated samples. If summary = 5, the function provides all simulated observations in the n_sim * k matrix.

Value

total number of colony forming units in the single mixing plan

Details

Let \(N'\) be the number of colony-forming units in the mixed sample which is produced by mixing of \(k\) primary samples and \(N' = \sum N_i\). To more details, please refer the details section of compare_mixing_3.

References

Nauta, M.J., 2005. Microbiological risk assessment models for partitioning and mixing during food handling. International Journal of Food Microbiology 100, 311-322.

See also

Examples

mu <- 100
sigma <- 0.8
alpha <- 0.1
k <- 30
n_sim <- 20000
sim_single(mu, sigma, alpha, k, distribution = "Poisson lognormal-Type B", n_sim, summary = 2)
#>  [1]  0  2  0 61  0  1  0  0 18  2  0  0  0  0  9 12  0  1  0  0  0  0  0  1  0 0  0  0  0  0