The total number of colony-forming units in the mixed sample by simulation result (in the single mixing plan).

This function calculates the resulting total number of colony forming units in the mixed sample in the single mixing plan. (to be finished later on)

sim_single(n_iter, mu, sigma, b, k, distribution, summary = FALSE)

Arguments

n_iter
number of iterations

mu
average number of colony-forming units in mixed sample which is in logarithmic scale if we use a lognormal distribution

sigma
log standard deviation of the colony-forming units in the mixed sample

b
concentration parameter

k
number of small portions/ primary samples

distribution
what suitable distribution type we have employed for simulation such as "Poisson-Type A" or "Poisson-Type B" or "Lognormal-Type A" or "Lognormal-Type B"

summary
if we need to get the mean and standard deviation of simulated \(N'\), use summary = TRUE ( default summary =FALSE).

Value

total number of colony forming units in the single mixing plan

Details

Let \(N'\) be the number of colony-forming units in the mixed sample which is produced by mixing of \(k\) primary samples and \(N' = \sum N_i\).To more details, please refer the details section of compare_mixing. (to be finished later on)

References

Nauta, M.J., 2005. Microbiological risk assessment models for partitioning and mixing during food handling. International Journal of Food Microbiology 100, 311-322.

See also

Examples

n_iter <- 200000
mu <- 100
sigma <- 0.8
b <- 0.1
k <- 10
sim_single( n_iter, mu, sigma, b, k, distribution = "Poisson-Type A", summary = TRUE)

#> Calculation took 40.76 seconds.

#>     mean_N   St.D_N
#> 1 99.99069 10.00949